# backend/tests/contracts/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "Body_upload_file_api_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanAdminPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PlanAdminPage",
        "type": "object"
      },
      "PlanIn": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Code",
            "type": "string"
          },
          "label_en": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label En",
            "type": "string"
          },
          "label_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label Id",
            "type": "string"
          },
          "months": {
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Months",
            "type": "integer"
          },
          "order": {
            "default": 1,
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Order",
            "type": "integer"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "label_id",
          "label_en",
          "months",
          "price"
        ],
        "title": "PlanIn",
        "type": "object"
      },
      "PlanOut": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label_en": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label En",
            "type": "string"
          },
          "label_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label Id",
            "type": "string"
          },
          "months": {
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Months",
            "type": "integer"
          },
          "order": {
            "default": 1,
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Order",
            "type": "integer"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "integer"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label_id",
          "label_en",
          "months",
          "price",
          "id"
        ],
        "title": "PlanOut",
        "type": "object"
      },
...
      "SnapTokenIn": {
        "properties": {
          "partner_id": {
            "title": "Partner Id",
            "type": "string"
          },
          "plan_code": {
            "title": "Plan Code",
            "type": "string"
          }
        },
        "required": [
          "partner_id",
          "plan_code"
        ],
        "title": "SnapTokenIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Explore Wisata Sumut API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/premium/plans": {
      "get": {
        "operationId": "list_admin_plans_api_admin_premium_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "active",
                "inactive"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "order",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAdminPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Admin Plans"
      },
      "post": {
        "operationId": "create_plan_api_admin_premium_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan"
      }
    },
    "/api/admin/premium/plans/{plan_id}": {
      "delete": {
        "operationId": "delete_plan_api_admin_premium_plans__plan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Plan"
      },
      "put": {
        "operationId": "update_plan_api_admin_premium_plans__plan_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Plan"
      }
    },
...
    "/api/files/{path}": {
      "get": {
        "operationId": "serve_file_api_files__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Serve File"
      }
    },
...
    "/api/mitra/partners/{partner_id}/payments": {
      "get": {
        "operationId": "list_my_partner_payments_api_mitra_partners__partner_id__payments_get",
        "parameters": [
          {
            "in": "path",
            "name": "partner_id",
            "required": true,
            "schema": {
              "title": "Partner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Partner Payments"
      }
    },
    "/api/mitra/partners/{partner_id}/payments/{order_id}/retry": {
      "post": {
        "operationId": "retry_my_partner_payment_api_mitra_partners__partner_id__payments__order_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "partner_id",
            "required": true,
            "schema": {
              "title": "Partner Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry My Partner Payment"
      }
    },
...
    "/api/payments/snap-token": {
      "post": {
        "operationId": "create_snap_token_api_payments_snap_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Snap Token"
      }
    },
    "/api/payments/{order_id}/status": {
      "get": {
        "operationId": "payment_status_api_payments__order_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Status"
      }
    },
...
    "/api/upload": {
      "post": {
        "operationId": "upload_file_api_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File"
      }
    },
...
  }
}